Move setup related scripts into scripts/ directory
Cleaning up root a bit.
