+++ scripts/checknix.py
